move feathr configuration
